+++ compiler-tests/src/test/java/dev/zacsweers/metro/compiler/OmitRedundantMirrorsIrOnlyClassesBoxTestGenerated.java
@@ -3106,6 +3106,12 @@ public class OmitRedundantMirrorsIrOnlyClassesBoxTestGenerated extends AbstractO
       }
 
       @Test
+      @TestMetadata("ClassMapKeyWrappedInterop.kt")
+      public void testClassMapKeyWrappedInterop() {
+        run("ClassMapKeyWrappedInterop.kt");
+      }
+
+      @Test
       @TestMetadata("KClassMapKeyInterop.kt")
       public void testKClassMapKeyInterop() {
         run("KClassMapKeyInterop.kt");